requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
astroid==3.0.3
async-timeout==4.0.3
asyncpg==0.29.0
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.8
exceptiongroup==1.2.0
fastapi==0.109.2
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.5
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
pylint==3.0.3
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
zipp==3.17.0
